fix jwt reissue and readd searchFeed

+++ src/main/java/com/meokplaylist/api/dto/socialInteraction/SaveFavoritePlaceDto.java
@@ -3,10 +3,10 @@ package com.meokplaylist.api.dto.socialInteraction;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-@Data
+@Data  
 @AllArgsConstructor
 public class SaveFavoritePlaceDto {
-    private Long x;
-    private Long y;
+    private double x;
+    private double y;
 
 }
