+++ src/main/java/com/meokplaylist/domain/repository/place/PlacesRepository.java
@@ -4,5 +4,5 @@ import com.meokplaylist.infra.place.Places;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PlacesRepository extends JpaRepository<Places, Long> {
-    Places findByLatitudeAndLongitude(Long x, Long y);
+    Places findByLatitudeAndLongitude(Double x, Double y);
 }
